requirements.txt
python-dateutil
netaddr
djangorestframework
coreapi
django_select2

django-dept-directory

